move fips functions to separate class
